refactor: extract account settings UI into a dedicated component and move auth persistence logic to a custom hook.
